mongo db

+++ CMS.ViewModel/Base/PagingBaseViewModel.cs
@@ -9,7 +9,7 @@ namespace CMS.ViewModel.Base
         public SortOrder Order { get; set; } = SortOrder.Ascending;
         public string OrderBy { get; set; } = "Id";
 
-        public int TotalPage { get; set; }
-        public int Total { get; set; }
+        public long TotalPage { get; set; }
+        public long Total { get; set; }
     }
 }
